backend/requirements.txt
fastapi>=0.110.0
uvicorn==0.25.0
motor>=3.4.0
pymongo>=4.6.0

python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1

python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic>=2.6.0
email-validator>=2.1.0.post1
